refactor: remove api response interface to _shared folder on api
